Make 페이지 관련 코드 별도로 추출
