+++ Bible2PPT/Bible2PPT.csproj
@@ -92,9 +92,14 @@
     </Compile>
     <Compile Include="AppConfig.cs" />
     <Compile Include="MainForm.Helper.cs">
+      <DependentUpon>MainForm.cs</DependentUpon>
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="IndexKeyAttribute.cs" />
+    <Compile Include="MainForm.Make.cs">
+      <DependentUpon>MainForm.cs</DependentUpon>
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Include="MultiPanel.cs">
       <SubType>Component</SubType>
     </Compile>
